+++ week2/shopping list/cart.html
@@ -58,5 +58,7 @@
 
     </main>
 </body>
-<script src="cart.js"></script>
+<script type="module" src="cartJS/cartTable.js"></script>
+<script type="module" src="cartJS/checkItem.js"></script>
+
 </html>
