+++ week2/shopping list/index.html
@@ -67,5 +67,8 @@
         <h1>...이것은 실제 저의 찜목록들을 가져온 것입니다...</h1>
     </footer>
 </body>
-<script src="index.js"></script>
+<script type="module" src="homeJS/nav.js"></script>
+<script type="module" src="homeJS/banner.js"></script>
+<script type="module" src="homeJS/modalLeft.js"></script>
+<script type="module" src="homeJS/mainItem.js"></script>
 </html>
