+++ week2/shopping list/interface/shopList.js
@@ -1,0 +1,15 @@
+//shop items
+export const shopList = [
+    { src: "img/items/a1.jpg", sort:'악세서리', title: "강아지 폰케이스", price: 5900},
+    { src: "img/items/a2.jpg", sort:'악세서리', title: "쪼꼬미 미니 리본핀", price: 6900},
+    { src: "img/items/a3.jpg", sort:'악세서리', title: "행운의 네잎 클로버 지비츠", price: 12900},
+    { src: "img/items/s1.jpg", sort:'하의', title: "레거시 플리츠 스커트", price: 26040},
+    { src: "img/items/s2.jpg", sort:'하의', title: "리본 다트 밴딩 미니스커트", price: 13850},
+    { src: "img/items/s3.jpg", sort:'하의', title: "스위티 프릴 스커트", price: 16740},
+    { src: "img/items/t1.jpg", sort:'상의', title: "프릴 슈 블라우스", price: 22620},
+    { src: "img/items/t2.jpg", sort:'상의', title: "인디 잔꽃 원피스 세트", price: 37900},
+    { src: "img/items/t3.jpg", sort:'상의', title: "레이어드 뷔스티에 미니원피스", price: 16020},
+    { src: "img/items/t4.jpg", sort:'상의', title: "프린지 후크 트위드자켓", price: 38610},
+    { src: "img/items/t5.jpg", sort:'상의', title: "리본 오프숄더", price: 37640},
+];
+export const ITEMS_KEY = "storedItems";
